feat: deploy DB en heroku y ejecutando migraciones

+++ database/migrations/2022_04_10_164614_create_parties_table.php
@@ -16,7 +16,7 @@ class CreatePartiesTable extends Migration
         Schema::create('parties', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('GameID');
+            $table->unsignedBigInteger('GameID');
             $table->foreign('GameID')
             ->references('id')
             ->on('games')
@@ -24,7 +24,7 @@ class CreatePartiesTable extends Migration
             ->constrained('games')
             ->onUpdate('cascade')
             ->onDelete('cascade');
-            $table->unsignedInteger('OwnerID');
+            $table->unsignedBigInteger('OwnerID');
             $table->foreign('OwnerID')
             ->references('id')
             ->on('users')
